Feat: Differentiate location coordinates from measurement coordinates

+++ internal/locations/models/models.go
@@ -2,7 +2,7 @@ package models
 
 type ThingLocation struct {
 	ThingURN          string  `json:"thing_urn" db:"thing_urn"`
-	LocationID        int     `json:"location_id" db:"location_id"`
+	LocationID        int64   `json:"location_id" db:"location_id"`
 	LocationName      string  `json:"location_name" db:"location_name"`
 	LocationLatitude  float64 `json:"location_latitude" db:"location_latitude"`
 	LocationLongitude float64 `json:"location_longitude" db:"location_longitude"`
